+++ batchflow/research/research.py
@@ -4,21 +4,19 @@ import csv
 import itertools
 import subprocess
 import re
-import functools
-import multiprocess as mp
 import dill
+import multiprocess as mp
 import glob
 import tqdm
 from collections import OrderedDict
 
-import numpy as np
-import pandas as pd
 import warnings
 
 from .domain import Domain
-from .distributor import Distributor
+from .distributor import Distributor, DynamicQueue
 from .experiment import Experiment, Executor
-from .utils import create_logger, to_list
+from .results import ResearchResults
+from .utils import create_logger
 
 class Research:
     def __init__(self, name='research', domain=None, experiment=None, n_configs=None, n_reps=1, repeat_each=None):
@@ -47,6 +45,7 @@ class Research:
         self.n_reps = n_reps
         self.repeat_each = repeat_each
         self._env = dict()
+        self._results = None
 
     def add_instance(self, *args, **kwargs):
         self.experiment.add_instance(*args, **kwargs)
@@ -90,10 +89,6 @@ class Research:
         """
         self.domain.set_update(function, when, **kwargs)
         return self
-
-    def load_results(self, *args, **kwargs):
-        """ Load results of research as pandas.DataFrame or dict (see :meth:`~.Results.load`). """
-        return Results(self.name, *args, **kwargs)
 
     def attach_env_meta(self, **kwargs):
         """ Save the information about the current state of project repository: commit, diff, status and others.
@@ -175,16 +170,12 @@ class Research:
         return devices
 
     def create_research_folder(self):
-        if not os.path.exists(self.name):
-            os.makedirs(self.name)
-            for subfolder in ['configs', 'description', 'env', 'experiments']:
-                config_path = os.path.join(self.name, subfolder)
-                if not os.path.exists(config_path):
-                    os.makedirs(config_path)
-        else:
-            raise ValueError(
-                "Research with name '{}' already exists".format(self.name)
-            )
+        os.makedirs(self.name)
+        for subfolder in ['configs', 'description', 'env', 'experiments']:
+            config_path = os.path.join(self.name, subfolder)
+            if not os.path.exists(config_path):
+                os.makedirs(config_path)
+
 
     def run(self, name=None, workers=1, branches=1, n_iters=None, devices=None, executor_class=None,
             dump_results=True, parallel=True, executor_target='threads', loglevel='debug', bar=True):
@@ -235,6 +226,7 @@ class Research:
                 workers = [workers[0]]
 
         self.name = name or self.name
+
         self.workers = workers
         self.branches = branches
         self.n_iters = n_iters
@@ -246,16 +238,20 @@ class Research:
         self.loglevel = loglevel
         self.bar = bar
 
-        if self.dump_results:
-            self.create_research_folder()
-            self.experiment = self.experiment.dump()
-        self.attach_env_meta()
+        if dump_results and os.path.exists(self.name):
+            raise ValueError(f"Research with name '{self.name}' already exists")
 
         self.domain.set_iter_params(n_items=self.n_configs, n_reps=self.n_reps, repeat_each=self.repeat_each)
 
         if self.domain.size is None and (self.domain.update_func is None or self.domain.update_each == 'last'):
             warnings.warn("Research will be infinite because has infinite domain and hasn't domain updating",
                           stacklevel=2)
+
+        if self.dump_results:
+            self.create_research_folder()
+            self.experiment = self.experiment.dump() # add final dump of experiment results
+            self.dump_research()
+        self.attach_env_meta()
 
         self.create_logger()
 
@@ -265,8 +261,8 @@ class Research:
         self.tasks_queue = DynamicQueue(self.domain, self, n_branches)
         self.distributor = Distributor(self.tasks_queue, self)
 
-        self.results = ResearchResults(self.name, self.dump_results)
         self.monitor = ResearchMonitor(self, self.name, bar=self.bar) # process execution signals
+        self.results = ResearchResults(self.name, self.dump_results)
 
         self.monitor.start(self.dump_results)
         self.distributor.run()
@@ -279,6 +275,17 @@ class Research:
         path = os.path.join(self.name, 'research.log') if self.dump_results else None
 
         self.logger = create_logger(name, path, self.loglevel)
+
+    def dump_research(self):
+        with open(os.path.join(self.name, 'research.dill'), 'wb') as f:
+            dill.dump(self, f)
+
+    @classmethod
+    def load(cls, name):
+        with open(os.path.join(name, 'research.dill'), 'rb') as f:
+            research = dill.load(f)
+        research.results = ResearchResults(research.name, research.dump_results)
+        return research
 
     def __str__(self):
         spacing = ' ' * 4
@@ -298,82 +305,6 @@ class Research:
             repr += '\n'
 
         return repr
-
-class DynamicQueue:
-    """ Queue of tasks that can be changed depending on previous results. """
-    def __init__(self, domain, research, n_branches):
-        self._domain = domain
-        self.research = research
-        self.n_branches = n_branches
-
-        self.queue = mp.JoinableQueue()
-
-        self.configs_generated = 0
-        self.configs_remains = self._domain.size
-
-        self.finished_tasks = 0
-        self.tasks_in_queue = 0
-
-    @property
-    def domain(self):
-        if self._domain.size == 0:
-            domain = Domain({'repetition': [None]}) # the value of repetition will be rewritten
-            domain.set_iter_params(n_reps=self._domain.n_reps, produced=self._domain.n_produced)
-            self._domain = domain
-        return self._domain
-
-    def update_domain(self):
-        """ Update domain. """
-        new_domain = self.domain.update(self.configs_generated, self.research)
-        if new_domain is not None:
-            self._domain = new_domain
-            self.configs_remains = self._domain.size
-        return new_domain is not None
-
-    def next_tasks(self, n_tasks=1):
-        """ Get next `n_tasks` elements of queue. """
-        configs = []
-        for i in range(n_tasks):
-            branch_tasks = [] # TODO: rename it
-            try:
-                for _ in range(self.n_branches):
-                    branch_tasks.append(next(self.domain))
-                configs.append(branch_tasks)
-            except StopIteration:
-                if len(branch_tasks) > 0:
-                    configs.append(branch_tasks)
-                break
-        for i, executor_configs in enumerate(configs):
-            self.put((self.configs_generated + i, executor_configs))
-
-        n_configs = sum([len(item) for item in configs])
-
-        self.configs_generated += n_configs
-        self.configs_remains -= n_configs
-        self.tasks_in_queue += len(configs)
-        self.research.logger.info(f'Get {n_tasks} tasks with {n_configs} configs, remains {self.configs_remains}')
-
-        return n_configs
-
-    def stop_workers(self, n_workers):
-        """ Stop all workers by putting `None` task into queue. """
-        for _ in range(n_workers):
-            self.put(None)
-
-    def join(self):
-        self.queue.join()
-
-    def get(self):
-        return self.queue.get()
-
-    def put(self, *args, **kwargs):
-        return self.queue.put(*args, **kwargs)
-
-    def task_done(self):
-        return self.queue.task_done()
-
-    def empty(self):
-        return self.queue.empty()
 
 class ResearchMonitor:
     COLUMNS = ['time', 'task_idx', 'id', 'it', 'name', 'status', 'exception', 'worker', 'pid', 'worker_pid',
@@ -491,148 +422,3 @@ class ResearchMonitor:
     def stop(self):
         self.queue.put(None)
         self.stop_signal.get()
-
-class ResearchResults:
-    def __init__(self, name, dump_results):
-        self.name = name
-        self.dump_results = dump_results
-        self.results = mp.Manager().dict()
-        self.configs = mp.Manager().dict()
-
-    def put(self, id, results, config):
-        self.results[id] = results
-        self.configs[id] = config
-
-    def to_df(self, pivot=False, include_config=True, **kwargs):
-        df = []
-        for experiment_id in self.results:
-            experiment_df = []
-            for name in self.results[experiment_id]:
-                _df = {
-                    'id': experiment_id,
-                    'iteration': self.results[experiment_id][name].keys()
-                }
-                if pivot:
-                    _df[name] = self.results[experiment_id][name].values()
-                else:
-                    _df['name'] = name
-                    _df['value'] = self.results[experiment_id][name].values()
-                experiment_df += [pd.DataFrame(_df)]
-            if pivot and len(experiment_df) > 0:
-                experiment_df = [functools.reduce(functools.partial(pd.merge, on=['id', 'iteration']), experiment_df)]
-            df += experiment_df
-        res = pd.concat(df) if len(df) > 0 else pd.DataFrame()
-        if include_config and len(res) > 0:
-            res = pd.merge(res, self.configs_to_df(**kwargs), how='inner', on='id')
-        return res
-
-    def load_iteration_files(self, path, iteration):
-        filenames = glob.glob(os.path.join(path, '*'))
-        if iteration is None:
-            files_to_load = {int(os.path.basename(filename)): filename for filename in filenames}
-        else:
-            dumped_iteration = np.sort(np.array([int(os.path.basename(filename)) for filename in filenames]))
-            files_to_load = dict()
-            for it in iteration:
-                _it = dumped_iteration[np.argwhere(dumped_iteration >= it)[0, 0]]
-                files_to_load[_it] = os.path.join(path, str(_it))
-        files_to_load = OrderedDict(sorted(files_to_load.items()))
-        results = OrderedDict()
-        for filename in files_to_load.values():
-            with open(filename, 'rb') as f:
-                values = dill.load(f)
-                for it in values:
-                    if iteration is None or it in iteration:
-                        results[it] = values[it]
-        return results
-
-    def load(self, experiment_id=None, name=None, iteration=None, config=None, alias=None, domain=None, **kwargs):
-        experiment_id = experiment_id if experiment_id is None else to_list(experiment_id)
-        name = name if name is None else to_list(name)
-        iteration = iteration if iteration is None else to_list(iteration)
-
-        filtered_ids = self.filter_ids_by_configs(config, alias, domain, **kwargs)
-        experiment_id = np.intersect1d(experiment_id, filtered_ids) if experiment_id is not None else filtered_ids
-
-        if self.dump_results:
-            self.results = OrderedDict()
-            for path in glob.glob(os.path.join(self.name, 'experiments', '*', 'results', '*')):
-                path = os.path.normpath(path)
-                _experiment_id, _, _name = path.split(os.sep)[-3:]
-                if experiment_id is None or _experiment_id in experiment_id:
-                    if name is None or _name in name:
-                        if _experiment_id not in self.results:
-                            self.results[_experiment_id] = OrderedDict()
-                        experiment_results = self.results[_experiment_id]
-
-                        if _name not in experiment_results:
-                            experiment_results[_name] = OrderedDict()
-                        name_results = experiment_results[_name]
-                        new_values = self.load_iteration_files(path, iteration)
-                        experiment_results[_name] = OrderedDict([*name_results.items(), *new_values.items()])
-
-    def configs_to_df(self, use_alias=True, concat_config=False, remove_auxilary=True):
-        df = []
-        for experiment_id in self.configs:
-            config = self.configs[experiment_id]
-            if remove_auxilary:
-                for key in ['repetition', 'device', 'updates']:
-                    config.pop_config(key)
-            if use_alias:
-                if concat_config:
-                    config = {'config': config.alias(as_string=concat_config)}
-                else:
-                    config = config.alias()
-            else:
-                config = config.config()
-            df += [pd.DataFrame({'id': [experiment_id], **config})]
-        return pd.concat(df)
-
-    def filter_ids_by_configs(self, config=None, alias=None, domain=None, **kwargs):
-        """ Filter configs.
-
-        Parameters
-        ----------
-        repetition : int, optional
-            index of the repetition to load, by default None
-        experiment_id : str or list, optional
-            experiment id to load, by default None
-        configs : dict, optional
-            specify keys and corresponding values to load results, by default None
-        aliases : dict, optional
-            the same as `configs` but specify aliases of parameters, by default None
-
-        Returns
-        -------
-        list
-            filtered list on configs
-        """
-        if sum([domain is not None, config is not None, alias is not None]) > 1:
-            raise ValueError('Only one of `config`, `alias` and `domain` can be not None')
-        filtered_ids = []
-        if domain is not None:
-            for config in domain.iterator():
-                filtered_ids += self.filter_ids_by_configs(config=config.config())
-            return filtered_ids
-
-        if len(kwargs) > 0:
-            if config is not None:
-                config = {**config, **kwargs}
-            elif alias is not None:
-                alias = {**alias, **kwargs}
-            else:
-                config = kwargs
-
-        if config is None and alias is None:
-            return list(self.configs.keys())
-
-        for experiment_id, supconfig in self.configs.items():
-            if config is not None:
-                _config = supconfig.config()
-                if all(item in _config.items() for item in config.items()):
-                    filtered_ids += [experiment_id]
-            else:
-                _config = supconfig.alias()
-                if all(item in _config.items() for item in alias.items()):
-                    filtered_ids += [experiment_id]
-        return filtered_ids
